pyproject.toml
[project]
name = "vibeto-chatbot"
version = "0.1.0"
description = "Vibe-to-Spec Prompt Architect — Convert vague ideas into Agent-ready spec documents"
requires-python = ">=3.11"

dependencies = [
    "langgraph>=0.2",
    "langchain-google-genai>=2.0",
    "langchain-core>=0.3",
    "langsmith>=0.1",
    "pydantic>=2.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "streamlit>=1.40",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
